Refactoring CouchTrackerSyncTests
