+++ CouchTrackerSyncTests/SyncWatchedShows/SyncWatchedShowsTests.swift
@@ -1,22 +1,15 @@
 import XCTest
-@testable import CouchTrackerSync
 import Moya
-
 import TraktSwift
 import TraktSwiftTestable
-
 import RxSwift
 import RxTest
-
 import Nimble
 import RxNimble
 
-private final class TestableClass: NSObject {}
-
-private let bundle = Bundle(for: TestableClass.self)
+@testable import CouchTrackerSync
 
 final class SyncWatchedShowsTests: XCTestCase {
-  private let userDefaultsMock = UserDefaults(suiteName: "SyncWatchedShowsTestsUserDefaults")!
   private var originalTrakt: TraktProvider!
   private var scheduler: TestScheduler!
   private var disposeBag: DisposeBag!
@@ -38,12 +31,11 @@ final class SyncWatchedShowsTests: XCTestCase {
 
   func testSyncWatchedShows() throws {
     let data = contentsOf(file: "SyncWatchedShowsSuccess")
-
     let responseStubbed = EndpointSampleResponse.networkResponse(200, data)
+    let trakt: SyncTestableTrakt<Sync> = makeTestableTrakt(responseStubbed)
+    Current.trakt = trakt
 
     let expectedShows = try JSONDecoder().decode([BaseShow].self, from: data)
-
-    Current.trakt = makeTrakt(responseStubbed)
 
     expect(syncWatchedShows())
       .events(scheduler: scheduler, disposeBag: disposeBag)
@@ -51,28 +43,5 @@ final class SyncWatchedShowsTests: XCTestCase {
         Recorded.next(0, expectedShows),
         Recorded.completed(0)
       ]))
-  }
-
-  private func contentsOf(file named: String) -> Data {
-    return bundle.url(forResource: named, withExtension: "json").map { try! Data(contentsOf: $0) }!
-  }
-
-  private func makeTrakt(_ sampleResponse: EndpointSampleResponse) -> SyncTestableTrakt<Sync> {
-    let builder = TraktBuilder {
-      $0.clientId = "clientIdMock"
-      $0.clientSecret = "clientSecretMock"
-      $0.redirectURL = "https://google.com"
-      $0.userDefaults = userDefaultsMock
-    }
-
-    return SyncTestableTrakt(builder: builder, endpointClosure: { target in
-      Endpoint(
-        url: URL(target: target).absoluteString,
-        sampleResponseClosure: { sampleResponse },
-        method: target.method,
-        task: target.task,
-        httpHeaderFields: target.headers
-      )
-    })
   }
 }
